refactor(theme-classic): extract HomeBreadcrumbItem (#6953)
