+++ packages/docusaurus-theme-classic/src/theme-classic.d.ts
@@ -1510,3 +1510,7 @@ declare module '@theme/prism-include-languages' {
     PrismObject: typeof PrismNamespace,
   ): void;
 }
+
+declare module '@theme/DocBreadcrumbs/HomeBreadcrumbItem' {
+  export default function HomeBreadcrumbItem(): JSX.Element;
+}
