+++ packages/docusaurus-theme-classic/src/theme/DocBreadcrumbs/HomeBreadcrumbItem/styles.module.css
@@ -5,7 +5,10 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-.breadcrumbsContainer {
-  --ifm-breadcrumb-size-multiplier: 0.8;
-  margin-bottom: 0.8rem;
+.breadcrumbHomeIcon {
+  position: relative;
+  top: 1px;
+  vertical-align: top;
+  height: 1.1rem;
+  width: 1.1rem;
 }
